fingbugs ICAST_INTEGER_MULTIPLY_CAST_TO_LONG
git-svn-id: https://svn.code.sf.net/p/autoplot/code/autoplot/trunk@27092 0c594bb0-8543-0410-a25e-d8ec440261db

+++ BinaryDataSource/src/org/autoplot/binarydatasource/BinaryDataSource.java
@@ -344,7 +344,7 @@ public class BinaryDataSource extends AbstractDataSource {
                 MutablePropertyDataSet dep0ds= new AbstractRank1DataSet(frecCount) {
                     @Override
                     public double value(int i) {
-                        return offset + finalRecOffset + i * finalRecSizeBytes;
+                        return offset + finalRecOffset + ((long)i) * finalRecSizeBytes;
                     }
                 };
                 dep0ds.putProperty( QDataSet.CADENCE, DataSetUtil.asDataSet((double)recSizeBits/8) );
